test(e2e): fold 2304 template tests into test_post_test_plan, satisfy path-mirror
